Add comprehensive functional tests for all 20 cases
- Fixed scoring granularity: use float scores instead of rounded integers
  for better discrimination between models
- Added behavioral tests for 15 missing cases (04-20):
  * case_04_notes: 9 tests for note creation, persistence, search, edit
  * case_05_weather: 10 tests for city display, weather data, responsiveness
  * case_07_stopwatch: 10 tests for timer modes, lap, reset functionality
  * case_08_typing: 10 tests for WPM, accuracy, typing mechanics
  * case_09_expenses: 10 tests for expense tracking, categories, charts
  * case_11_palette: 10 tests for color generation, lock, copy features
  * case_12_habits: 10 tests for habit tracking, streaks, persistence
  * case_13_password: 10 tests for password generation, options, strength
  * case_14_units: 10 tests for unit conversion, categories, live update
  * case_15_markdown: 10 tests for markdown rendering, preview, toolbar
  * case_16_repostats: 8 tests for git stats Python script validation
  * case_17_legal: 10 tests for legal case search, notes, metadata
  * case_18_slides: 7 tests for slide summary Python script validation
  * case_19_control_center: 10 tests for dashboard, views, glassmorphism
  * case_20_log_analytics: 9 tests for log parsing, HTTP server, charts

- Tightened existing Tier 1 tests to be more discriminating:
  * case_01_pomodoro: Added pause, reset, countdown, session counter tests
  * case_02_quiz: Improved quiz navigation and feedback tests
  * case_03_calculator: Added subtraction, division, decimal, chained ops tests

- Tightened Tier 2 tests to verify actual functionality:
  * case_21_spreadsheet: Added SUM function, cell reference, keyboard nav tests
  * case_22_flowchart: Added shape types, drag, connect, undo tests
  * case_23_richtext: Added Ctrl+B/I shortcuts, font size, alignment tests
  * case_24_filebrowser: Added specific folder/file tests, breadcrumb, dates
  * case_25_dataviz: Added chart type verification, legend, tooltip tests

These improvements should provide much better discrimination between
model capabilities on the benchmark.

+++ vibe_eval/scoring/aggregator.py
@@ -17,7 +17,7 @@ from ..judge.absolute import AbsoluteScore
 class DimensionResult:
     """Result for a single scoring dimension."""
     name: str
-    score: int  # 0-10
+    score: float  # 0-10 (float for granularity)
     weight: float  # Weight as decimal (e.g., 0.20 for 20%)
     source: str  # "auto", "judge", "static", or "combined"
     reason: str = ""
